+++ built_value_generator/tools/analyzer_plugin/bin/plugin.dart
@@ -1,0 +1,7 @@
+import 'dart:isolate';
+
+import 'package:built_value_generator/plugin_starter.dart';
+
+void main(List<String> args, SendPort sendPort) {
+  start(args, sendPort);
+}